% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chessgames.R
\docType{data}
\name{DeepBlueKasparov}
\alias{DeepBlueKasparov}
\title{Deep Blue versus Garry Kasparov (1997)}
\format{
An object of class \code{data.frame} with 37 rows and 2 columns.
}
\usage{
DeepBlueKasparov
}
\description{
Deep Blue versus Garry Kasparov (1997)
}
\keyword{datasets}
