% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chessgames.R
\docType{data}
\name{EinsteinOppenheimer}
\alias{EinsteinOppenheimer}
\title{Einstein versus Oppenheimer}
\format{
An object of class \code{data.frame} with 24 rows and 2 columns.
}
\usage{
EinsteinOppenheimer
}
\description{
Einstein versus Oppenheimer
}
\keyword{datasets}
