% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chessgames.R
\docType{data}
\name{ScachsDamor}
\alias{ScachsDamor}
\title{Scachs d'amor}
\format{
An object of class \code{data.frame} with 21 rows and 2 columns.
}
\usage{
ScachsDamor
}
\description{
Scachs d'amor
}
\keyword{datasets}
